Introduction of sample 3. Navigation between screens and passing parameters.
This sample has a problem... I am continuously navigating popping most stuff into the stack, but it would be awesome to know how to dequeue everything or how to make it going back but updating the store.

Things introduced that might be useful later on:
- Guid implementation.
- Resources and Controls (just a button at the moment)

Starting to find lots of samples online using redux, but will investigate a bit more about the navigation issue before jumping into redux.