Apparently calling an async result that might take a while to comeback can cause some bad karma.
The UI warning was telling me I had to call setState while the component was mounted, but I was calling
the async method in my componentDidMount method, what is wrong? Uh... **async** method.. ding it!

If I track the state of the mount and unmount and I prevent this issue.

I will have to comeback in the future to do more redux so I can understand properly how to tackle this
in a more redux(ish) way.